fix(webapps/assembly): Fix selectPDStatistics Query
Context: The query selectPDStatistics calculates the total processInstances wrong.
Used-by: ProcessDefinitionRestService#getStatisticsCount

Why: The group by incidentType part of the query should not be used for counting the total process instances.

Tests: The commit adds a test which expects the correct statistics returned when process instances related to incidents with different types are used.

Co-authored-by: Petros Savvidis <petros.savvidis@camunda.com>
Related to https://jira.camunda.com/browse/SUPPORT-24279